refactor(jwtAnyOne): change jwtAnyOne name to jwtAnyAuthorized
